java

+++ cracking_the_coding_interview/abba/ch01/src/main/java/org/example/IsUnique.java
@@ -1,17 +1,17 @@
 package org.example;
 class IsUnique {
     public boolean isUnique(String str) {
-        int bit = 0;
+        long bit = 0;
 
         if (str == null || str.isEmpty()) {
             return true;
         }
         for (int i = 0; i <  str.length(); i++ ) {
             int c = str.charAt(i) - 'a';
-            if ((bit & (1 << c)) > 0) {
+            if ((bit & (1L << c)) > 0) {
                 return false;
             }
-            bit |= (1 << c);
+            bit |= (1L << c);
         }
         return true;
     }
